Split the runner state parts into separate mods and restructure its internals
