+++ crates/lib/runner-state/src/lib.rs
@@ -1,11 +1,16 @@
+mod collect_value_sources;
+mod error;
+mod execution;
+mod max_nodes;
+mod resolve_value_tree;
 mod state;
 mod util;
-mod value {
-    pub mod resolver;
-    pub mod source_collector;
-}
+mod value;
 
+pub use self::error::*;
+pub use self::execution::*;
 pub use self::state::*;
+pub use self::value::*;
 
-use self::value::resolver::resolve_value_tree;
-use self::value::source_collector::collect_value_sources;
+use self::collect_value_sources::collect_value_sources;
+use self::resolve_value_tree::resolve_value_tree;
