+++ crates/lib/runner-state/src/resolve_value_tree.rs
@@ -2,12 +2,7 @@
 
 use std::collections::{HashMap, HashSet};
 
-use crate::ValueExpr;
-
-use crate::state::{
-    ActionCallSpec, BinaryOpValue, DictEntryValue, DictValue, DotValue, FunctionCallValue,
-    IndexValue, ListValue, SpreadValue, UnaryOpValue,
-};
+use crate::value::*;
 
 /// Resolve variables inside a ValueExpr tree without executing actions.
 ///
